services/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
pwdlib[argon2,bcrypt]
python-multipart
python-jose[cryptography]
python-multipart